@@ -14,18 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-include_once "templates/base.php";
-echo pageHeader("Simple API Access");
 
-/************************************************
-  Make a simple API request using a key. In this
-  example we're not making a request as a
-  specific user, but simply indicating that the
-  request comes from our application, and hence
-  should use our quota, which is higher than the
-  anonymous quota (which is limited per IP).
- ************************************************/
-require_once realpath(dirname(__FILE__) . '/../src/Google/autoload.php');
+include_once __DIR__ . '/../vendor/autoload.php';
+include_once "templates/base.php";
+
+echo pageHeader("Simple API Access");
 
 /************************************************
   We create the client and set the simple API
@@ -35,9 +28,9 @@ require_once realpath(dirname(__FILE__) . '/../src/Google/autoload.php');
  ************************************************/
 $client = new Google_Client();
 $client->setApplicationName("Client_Library_Examples");
-$apiKey = "<YOUR_API_KEY>"; // Change this line.
-// Warn if the API key isn't changed.
-if (strpos($apiKey, "<") !== false) {
+
+// Warn if the API key isn't set.
+if (!$apiKey = getApiKey()) {
   echo missingApiKeyWarning();
   exit;
 }
