@@ -15,9 +15,8 @@
  * limitations under the License.
  */
 
-if (!class_exists('Google_Client')) {
-  require_once dirname(__FILE__) . '/../autoload.php';
-}
+use Google\Auth\CacheInterface;
+use Psr\Log\LoggerInterface;
 
 /*
  * This class implements a basic on disk storage. While that does
@@ -27,21 +26,21 @@ if (!class_exists('Google_Client')) {
  *
  * @author Chris Chabot <chabotc@google.com>
  */
-class Google_Cache_File extends Google_Cache_Abstract
+class Google_Cache_File implements CacheInterface
 {
   const MAX_LOCK_RETRIES = 10;
   private $path;
   private $fh;
 
   /**
-   * @var Google_Client the current client
+   * @var use Psr\Log\LoggerInterface logger
    */
-  private $client;
+  private $logger;
 
-  public function __construct(Google_Client $client)
+  public function __construct($path, LoggerInterface $logger = null)
   {
-    $this->client = $client;
-    $this->path = $this->client->getClassConfig($this, 'directory');
+    $this->path = $path;
+    $this->logger = $logger;
   }
 
   public function get($key, $expiration = false)
@@ -50,7 +49,8 @@ class Google_Cache_File extends Google_Cache_Abstract
     $data = false;
 
     if (!file_exists($storageFile)) {
-      $this->client->getLogger()->debug(
+      $this->log(
+          'debug',
           'File cache miss',
           array('key' => $key, 'file' => $storageFile)
       );
@@ -60,7 +60,8 @@ class Google_Cache_File extends Google_Cache_Abstract
     if ($expiration) {
       $mtime = filemtime($storageFile);
       if ((time() - $mtime) >= $expiration) {
-        $this->client->getLogger()->debug(
+        $this->log(
+            'debug',
             'File cache miss (expired)',
             array('key' => $key, 'file' => $storageFile)
         );
@@ -74,7 +75,8 @@ class Google_Cache_File extends Google_Cache_Abstract
         $data = fread($this->fh, filesize($storageFile));
         $data =  unserialize($data);
       } else {
-        $this->client->getLogger()->debug(
+        $this->log(
+            'debug',
             'Cache file was empty',
             array('file' => $storageFile)
         );
@@ -82,7 +84,8 @@ class Google_Cache_File extends Google_Cache_Abstract
       $this->unlock($storageFile);
     }
 
-    $this->client->getLogger()->debug(
+    $this->log(
+        'debug',
         'File cache hit',
         array('key' => $key, 'file' => $storageFile, 'var' => $data)
     );
@@ -100,12 +103,14 @@ class Google_Cache_File extends Google_Cache_Abstract
       $result = fwrite($this->fh, $data);
       $this->unlock($storageFile);
 
-      $this->client->getLogger()->debug(
+      $this->log(
+          'debug',
           'File cache set',
           array('key' => $key, 'file' => $storageFile, 'var' => $value)
       );
     } else {
-      $this->client->getLogger()->notice(
+      $this->log(
+          'notice',
           'File cache set failed',
           array('key' => $key, 'file' => $storageFile)
       );
@@ -116,14 +121,16 @@ class Google_Cache_File extends Google_Cache_Abstract
   {
     $file = $this->getCacheFile($key);
     if (file_exists($file) && !unlink($file)) {
-      $this->client->getLogger()->error(
+      $this->log(
+          'error',
           'File cache delete failed',
           array('key' => $key, 'file' => $file)
       );
       throw new Google_Cache_Exception("Cache file could not be deleted");
     }
 
-    $this->client->getLogger()->debug(
+    $this->log(
+        'debug',
         'File cache delete',
         array('key' => $key, 'file' => $file)
     );
@@ -147,7 +154,8 @@ class Google_Cache_File extends Google_Cache_Abstract
     $storageDir = $this->path . '/' . substr(md5($file), 0, 2);
     if ($forWrite && ! is_dir($storageDir)) {
       if (! mkdir($storageDir, 0700, true)) {
-        $this->client->getLogger()->error(
+        $this->log(
+            'error',
             'File cache creation failed',
             array('dir' => $storageDir)
         );
@@ -166,7 +174,8 @@ class Google_Cache_File extends Google_Cache_Abstract
   {
     $rc = $this->acquireLock(LOCK_EX, $storageFile);
     if (!$rc) {
-      $this->client->getLogger()->notice(
+      $this->log(
+          'notice',
           'File cache write lock failed',
           array('file' => $storageFile)
       );
@@ -180,7 +189,8 @@ class Google_Cache_File extends Google_Cache_Abstract
     $mode = $type == LOCK_EX ? "w" : "r";
     $this->fh = fopen($storageFile, $mode);
     if (!$this->fh) {
-      $this->client->getLogger()->error(
+      $this->log(
+          'error',
           'Failed to open file during lock acquisition',
           array('file' => $storageFile)
       );
@@ -204,6 +214,13 @@ class Google_Cache_File extends Google_Cache_Abstract
   {
     if ($this->fh) {
       flock($this->fh, LOCK_UN);
+    }
+  }
+
+  private function log($level, $message, $context = array())
+  {
+    if ($this->logger) {
+      $this->logger->log($level, $message, $context);
     }
   }
 }
