@@ -15,9 +15,9 @@
  * limitations under the License.
  */
 
-if (!class_exists('Google_Client')) {
-  require_once dirname(__FILE__) . '/../autoload.php';
-}
+use GuzzleHttp\Message\Request;
+use GuzzleHttp\Stream\Stream;
+use GuzzleHttp\Url;
 
 /**
  * Manage large file uploads, which may be media but can be any type
@@ -74,27 +74,20 @@ class Google_Http_MediaFileUpload
    */
   public function __construct(
       Google_Client $client,
-      Google_Http_Request $request,
+      $request,
       $mimeType,
       $data,
       $resumable = false,
-      $chunkSize = false,
-      $boundary = false
+      $chunkSize = false
   ) {
     $this->client = $client;
     $this->request = $request;
     $this->mimeType = $mimeType;
     $this->data = $data;
-    $this->size = strlen($this->data);
     $this->resumable = $resumable;
-    if (!$chunkSize) {
-      $chunkSize = 256 * 1024;
-    }
     $this->chunkSize = $chunkSize;
     $this->progress = 0;
-    $this->boundary = $boundary;
 
-    // Process Media Request
     $this->process();
   }
 
@@ -117,6 +110,36 @@ class Google_Http_MediaFileUpload
   }
 
   /**
+   * Send the next part of the file to upload.
+   * @param [$chunk] the next set of bytes to send. If false will used $data passed
+   * at construct time.
+   */
+  public function nextChunk($chunk = false)
+  {
+    $resumeUri = $this->getResumeUri();
+
+    if (false == $chunk) {
+      $chunk = substr($this->data, $this->progress, $this->chunkSize);
+    }
+
+    $lastBytePos = $this->progress + strlen($chunk) - 1;
+    $headers = array(
+      'content-range' => "bytes $this->progress-$lastBytePos/$this->size",
+      'content-length' => strlen($chunk),
+      'expect' => '',
+    );
+
+    $request = new Request(
+        'PUT',
+        $resumeUri,
+        $headers,
+        Stream::factory($chunk)
+    );
+
+    return $this->makePutRequest($request);
+  }
+
+  /**
    * Return the HTTP result code from the last call made.
    * @return int code
    */
@@ -134,66 +157,35 @@ class Google_Http_MediaFileUpload
   * @return false|mixed false when the upload is unfinished or the decoded http response
   *
   */
-  private function makePutRequest(Google_Http_Request $httpRequest)
+  private function makePutRequest(Request $request)
   {
-    if ($this->client->getClassConfig("Google_Http_Request", "enable_gzip_for_uploads")) {
-      $httpRequest->enableGzip();
-    } else {
-      $httpRequest->disableGzip();
-    }
+    $http = $this->client->getHttpClient();
+    $response = $http->send($request);
+    $this->httpResultCode = $response->getStatusCode();
 
-    $response = $this->client->getIo()->makeRequest($httpRequest);
-    $response->setExpectedClass($this->request->getExpectedClass());
-    $code = $response->getResponseHttpCode();
-    $this->httpResultCode = $code;
-
-    if (308 == $code) {
+    if (308 == $this->httpResultCode) {
       // Track the amount uploaded.
-      $range = explode('-', $response->getResponseHeader('range'));
+      $range = explode('-', $response->getHeader('range'));
       $this->progress = $range[1] + 1;
 
       // Allow for changing upload URLs.
-      $location = $response->getResponseHeader('location');
+      $location = $response->getHeader('location');
       if ($location) {
         $this->resumeUri = $location;
       }
 
       // No problems, but upload not complete.
       return false;
-    } else {
-      return Google_Http_REST::decodeHttpResponse($response, $this->client);
-    }
-  }
-
-  /**
-   * Send the next part of the file to upload.
-   * @param [$chunk] the next set of bytes to send. If false will used $data passed
-   * at construct time.
-   */
-  public function nextChunk($chunk = false)
-  {
-    if (false == $this->resumeUri) {
-      $this->resumeUri = $this->fetchResumeUri();
     }
 
-    if (false == $chunk) {
-      $chunk = substr($this->data, $this->progress, $this->chunkSize);
-    }
-    $lastBytePos = $this->progress + strlen($chunk) - 1;
-    $headers = array(
-      'content-range' => "bytes $this->progress-$lastBytePos/$this->size",
-      'content-type' => $this->request->getRequestHeader('content-type'),
-      'content-length' => $this->chunkSize,
-      'expect' => '',
-    );
+    $result = $response->json();
+    $expectedClass = $this->request->getHeader('X-Php-Expected-Class');
 
-    $httpRequest = new Google_Http_Request(
-        $this->resumeUri,
-        'PUT',
-        $headers,
-        $chunk
-    );
-    return $this->makePutRequest($httpRequest);
+    if ($expectedClass) {
+      $result = new $expectedClass($result);
+    }
+
+    return $result;
   }
 
   /**
@@ -212,6 +204,7 @@ class Google_Http_MediaFileUpload
          'PUT',
          $headers
      );
+
      return $this->makePutRequest($httpRequest);
   }
 
@@ -221,18 +214,19 @@ class Google_Http_MediaFileUpload
    */
   private function process()
   {
-    $postBody = false;
+    $this->transformToUploadUrl();
+
+    $postBody = '';
     $contentType = false;
 
-    $meta = $this->request->getPostBody();
+    $meta = (string) $this->request->getBody();
     $meta = is_string($meta) ? json_decode($meta, true) : $meta;
 
     $uploadType = $this->getUploadType($meta);
-    $this->request->setQueryParam('uploadType', $uploadType);
-    $this->transformToUploadUrl();
+    $this->request->getQuery()->set('uploadType', $uploadType);
     $mimeType = $this->mimeType ?
         $this->mimeType :
-        $this->request->getRequestHeader('content-type');
+        $this->request->getHeader('content-type');
 
     if (self::UPLOAD_RESUMABLE_TYPE == $uploadType) {
       $contentType = $mimeType;
@@ -256,18 +250,13 @@ class Google_Http_MediaFileUpload
       $postBody = $related;
     }
 
-    $this->request->setPostBody($postBody);
+    $this->request->setBody(Stream::factory($postBody));
 
     if (isset($contentType) && $contentType) {
-      $contentTypeHeader['content-type'] = $contentType;
-      $this->request->setRequestHeaders($contentTypeHeader);
+      $this->request->setHeader('content-type', $contentType);
     }
-  }
 
-  private function transformToUploadUrl()
-  {
-    $base = $this->request->getBaseComponent();
-    $this->request->setBaseComponent($base . '/upload');
+    return $this->request;
   }
 
   /**
@@ -294,44 +283,63 @@ class Google_Http_MediaFileUpload
 
   public function getResumeUri()
   {
-    return ( $this->resumeUri !== null ? $this->resumeUri : $this->fetchResumeUri() );
+    if (is_null($this->resumeUri)) {
+      $this->resumeUri = $this->fetchResumeUri();
+    }
+
+    return $this->resumeUri;
   }
 
   private function fetchResumeUri()
   {
     $result = null;
-    $body = $this->request->getPostBody();
+    $body = $this->request->getBody();
     if ($body) {
       $headers = array(
         'content-type' => 'application/json; charset=UTF-8',
-        'content-length' => Google_Utils::getStrLen($body),
+        'content-length' => $body->getSize(),
         'x-upload-content-type' => $this->mimeType,
         'x-upload-content-length' => $this->size,
         'expect' => '',
       );
-      $this->request->setRequestHeaders($headers);
+      foreach ($headers as $key => $value) {
+        $this->request->setHeader($key, $value);
+      }
     }
 
-    $response = $this->client->getIo()->makeRequest($this->request);
-    $location = $response->getResponseHeader('location');
-    $code = $response->getResponseHttpCode();
+    $response = $this->client->getHttpClient()->send($this->request);
+    $location = $response->getHeader('location');
+    $code = $response->getStatusCode();
 
     if (200 == $code && true == $location) {
       return $location;
     }
+
     $message = $code;
-    $body = @json_decode($response->getResponseBody());
-    if (!empty($body->error->errors) ) {
+    $body = $response->json();
+    if (isset($body['error']['errors'])) {
       $message .= ': ';
-      foreach ($body->error->errors as $error) {
-        $message .= "{$error->domain}, {$error->message};";
+      foreach ($body['error']['errors'] as $error) {
+        $message .= "{$error[domain]}, {$error[message]};";
       }
       $message = rtrim($message, ';');
     }
 
     $error = "Failed to start the resumable upload (HTTP {$message})";
     $this->client->getLogger()->error($error);
+
     throw new Google_Exception($error);
+  }
+
+  private function transformToUploadUrl()
+  {
+    $parts = parse_url($this->request->getUrl());
+    if (!isset($parts['path'])) {
+      $parts['path'] = '';
+    }
+    $parts['path'] = '/upload' . $parts['path'];
+    $url = Url::fromString(Url::buildUrl($parts));
+    $this->request->setUrl($url);
   }
 
   public function setChunkSize($chunkSize)
